1. Store user statistics as json string 2. Handle null fuel efficiency data by ignoring it in score calculation 3. Handle null data by representing it with NaN 4. Use sorted list by DateTime for trip history 5. On login, sync trip history with Mojio servers 6. Reorganize core project files into separate folders 7. Other minor bug fixes

+++ Src/Mojio.Gamification.Android/Views/CircularIndicatorView.cs
@@ -25,7 +25,7 @@ namespace Mojio.Gamification.Android
 	
 		private float mIndicatorWidth = DEFAULT_INDICATOR_WIDTH;
 		private int mIndicatorDegrees;
-		private int mIndicatorValue = DEFAULT_START_VALUE;
+		private double mIndicatorValue = DEFAULT_START_VALUE;
 		private int mIndicatorRange = DEFAULT_MAX_VALUE;
 		private Color mIndicatorColor = Color.White;
 		private Color mTextColor = Color.White;
@@ -174,7 +174,7 @@ namespace Mojio.Gamification.Android
 			} else {
 				mIndicatorDegrees = MAX_DEGREES;
 			}
-			mIndicatorValue = (int) Math.Round (value, MidpointRounding.AwayFromZero);
+			mIndicatorValue = Math.Round (value, MidpointRounding.AwayFromZero);
 			if (mValueUpdateHandler != null) {
 				mValueUpdateHandler.SendEmptyMessage (0);
 			}
